feat: std.mem raw-pointer accessors + variadic externs + compiler fixes
Bundle of in-flight Aether language and stdlib changes driven by the
in-tree mquickjs port. Single commit covers thirteen prior incremental
ones — the work cleanly belongs together as one shippable unit.

New stdlib module for caller-allocated buffer access. Companion to
std.bytes (Aether-managed mutable buffers); std.mem is for buffers
Aether does NOT own (raw char* from C calls, malloc'd blocks, platform-
returned pointers). Module name is std.mem not std.ptr because `ptr` is
a reserved keyword for the type itself. First consumer: in-tree
mquickjs port — every UTF-8 decoder, bytecode interpreter, packed-format
parser, and JSValue tagged-pointer manipulation site.

Surface, grouped by purpose:

- Byte: get_byte / set_byte (NULL-defended, low-8-bit-masked setter)
- Pointer slot: get_ptr / set_ptr (memcpy-lowered for unaligned safety)
- Width-sized int: get_int / set_int (int32 slots), get_long / set_long
  (size_t / int64 slots)
- Typed array (14 accessors): get/set int8/uint8/int16/uint16/uint32/
  float32/float64 — all memcpy-lowered for unaligned packed wire formats
- Float bit-reinterpret: bits_of_float / float_from_bits (via union for
  exact IEEE-754 round-trip)
- Bit-counting: clz32 / clz64 (width-returning on zero input)
- Wide division: udiv64_32 (unsigned 64/32 — single insn on x86-64 / AArch64)
- Bare-fnptr invocation: call_fn3_int / call_fn3_void (per-signature
  shims; placeholder until typed function-pointer dispatch lands)

Five-case integration test covers round-trip, NULL defence, byte-range
masking, and the negative-zero IEEE-754 bit-pattern probe.

`extern foo(fmt: string, ...) -> int` lowered to a C variadic prototype.
Two spellings: named-param + trailing `...`, and bare `(...)`. Call
sites pass any number of trailing args literally.

Motivating use case: forwarding to printf-shaped C APIs (JS_ThrowSyntaxError,
fprintf, snprintf) without a hand-written C shim per call site. v1 is
extern-only — Aether-side function-DEFINING varargs (va_list consumption)
are out of scope and deferred to a future change.

Implementation:
- Lexer: new TOKEN_DOTDOTDOT (third-dot peek in the `..` range path)
- Parser: parse_extern_declaration tags `annotation = "varargs"`
- Codegen: appends `, ...` or bare `...` to the C prototype; coordinates
  with the strict-prototype emit from 0.153.0 so bare `(...)` does not
  collide with the no-arg `(void)` emission
- Type-checker: no change — extern call sites already skip arity
  enforcement

Four-case integration test covers named+`...`, bare `(...)`, end-to-end
printf execution, and a non-varargs sanity check.

- TYPE_FLOAT → C `double` consistently. Several emission sites were
  emitting `float` (32-bit) while others emitted `double` — silent ABI
  divergence between Aether modules and across FFI. Aether `float` has
  always been documented as IEEE-754 double-precision.

- ptr +/- int keeps `ptr` type. Was coercing to int and truncating the
  high 32 bits on 64-bit platforms; `arr_base + idx * sizeof_T` returned
  a half-pointer that subsequently page-faulted. Now `ptr + int → ptr`,
  `ptr - int → ptr`, with the symmetric `int + ptr` form accepted. Other
  shapes (ptr + ptr, ptr + long, ptr * int) still type-error.

- Struct-literal disambiguator respects `in_condition`. Pattern
  `if v < cmp {\n}` (empty body on next line) was being parsed as
  `if (v < (cmp { ... }))` — a struct-literal swallow of the if's body.
  Gated struct-literal lookahead inside condition contexts so the brace
  binds to the if-body. Regression test added.

The std.mem negative-zero probe now writes `0x8000000000000000`
directly, lifting a runtime-arithmetic workaround. The underlying lexer
fix is upstream's (PR #457 wide-literal pipeline) — our originally
included strtoull lexer patch was superseded and dropped during the
rebase onto 0.153.0.

Tests passing: tests/integration/std_mem_byte_access (16/16),
tests/integration/aether_extern_varargs (4/4), syntax / language /
regression smoke green.

+++ compiler/codegen/codegen.c
@@ -741,7 +741,8 @@ void emit_actor_to_header(CodeGenerator* gen, ASTNode* actor) {
                             const char* c_type = "int";
                             switch (field->type_kind) {
                                 case TYPE_INT: c_type = "int"; break;
-                                case TYPE_FLOAT: c_type = "float"; break;
+                                /* See get_c_type() — Aether `float` is always C `double`. */
+                                case TYPE_FLOAT: c_type = "double"; break;
                                 case TYPE_STRING: c_type = "const char*"; break;
                                 case TYPE_BOOL: c_type = "int"; break;
                                 case TYPE_BYTE: c_type = "unsigned char"; break;
@@ -931,7 +932,17 @@ const char* get_c_type(Type* type) {
         case TYPE_INT: return "int";
         case TYPE_INT64: return "int64_t";
         case TYPE_UINT64: return "uint64_t";
-        case TYPE_FLOAT: return "float";
+        /* Aether `float` lowers to C `double`. The naming is legacy
+         * (Aether predates having two FP types), but the storage and
+         * ABI have always been 8-byte IEEE-754 — local variables are
+         * stored as C-double via bare `1.0` literals (which are
+         * C-double, not C-float), and extern function signatures
+         * already emit `double` (see codegen_func.c:325). Struct
+         * fields and forward declarations were the holdouts emitting
+         * 4-byte C `float`, which created an ABI mismatch when an
+         * Aether-defined function was called from C with a `double`
+         * argument. Now consistent everywhere. */
+        case TYPE_FLOAT: return "double";
         case TYPE_BOOL: return "int";
         /* `unsigned char` (not `uint8_t`) so the compiler's strict-aliasing
          * exemption applies: code may legally read or write any other
@@ -1055,7 +1066,10 @@ static const char* get_abi_type(Type* type) {
         case TYPE_INT:    return "int32_t";
         case TYPE_INT64:  return "int64_t";
         case TYPE_UINT64: return "uint64_t";
-        case TYPE_FLOAT:  return "float";
+        /* Aether `float` is C `double` (8 bytes, binary64) — see
+         * get_c_type() for rationale. The public ABI (`aether_*`
+         * wrapper symbols emitted with --emit=lib) follows suit. */
+        case TYPE_FLOAT:  return "double";
         case TYPE_BOOL:   return "int32_t";
         case TYPE_BYTE:   return "unsigned char";
         case TYPE_STRING: return "const char*";
